feat/#279 : 댓글 입력창 재사용을 위한 컴포넌트 분리
